refactor(#3): transaction pool
refactor code for pool transaction add new file and remove default content in lib project
